made changes
made changes

+++ my-app/src/store/apiSlice.js
@@ -1,41 +1,61 @@
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
 
- 
 const baseURI = 'http://localhost:8080';
- 
+
 export const apiSlice = createApi({
-    baseQuery : fetchBaseQuery({ baseUrl : baseURI}),
-    endpoints : builder => ({
-        getCategories : builder.query({
-            query: () => '/v1/category',
-            providesTags: ['categories']
-        }),
- 
-        getLabels : builder.query({
-            query : () => '/v1/transaction/labels',
-            providesTags: ['transaction']
-        }),
- 
-        addTransaction : builder.mutation({
-            query : (initialTransaction) => ({
-                url: '/v1/transaction',
-                method: "POST",
-                body: initialTransaction
-            }),
-            invalidatesTags: ['transaction']
-        }),
- 
-        deleteTransaction : builder.mutation({
-            query : recordId => ({
-                url : `/v1/transaction/${recordId}`,
-                method : "DELETE",
-            }),
-            invalidatesTags: ['transaction']
-        })
- 
-    })
-})
- 
+  baseQuery: fetchBaseQuery({ baseUrl: baseURI }),
+  endpoints: (builder) => ({
+  
+    getLabels: builder.query({
+      query: () => {
+        const userId = localStorage.getItem('userId');
+
+        if (!userId) {
+          console.error('User ID not found in local storage.');
+          return;
+        }
+
+        return `/v1/transaction/list?userId=${userId}`;
+      },
+      providesTags: ['transaction'],
+    }),
+
+    addTransaction: builder.mutation({
+      query: (transactionData) => {
+        const userId = localStorage.getItem('userId');
+
+        if (!userId) {
+          console.error('User ID not found in local storage.');
+          return;
+        }
+
+        return {
+          url: `/v1/transaction?userId=${userId}`,
+          method: 'POST',
+          body: transactionData,
+        };
+      },
+      invalidatesTags: ['transaction'],
+    }),
+
+    deleteTransaction: builder.mutation({
+      query: (recordId) => {
+        const userId = localStorage.getItem('userId');
+
+        if (!userId) {
+          console.error('User ID not found in local storage.');
+          return;
+        }
+
+        return {
+          url: `/v1/transaction/${recordId}?userId=${userId}`,
+          method: 'DELETE',
+        };
+      },
+      invalidatesTags: ['transaction'],
+    }),
+
+  }),
+});
+
 export default apiSlice;
-
-
